From: Alan Webb <alan_webb@us.ibm.com>
  Modify dynamic texture handling so that an explicitly targetted
  face is not scheduled for immediate expiration. The requirement
  for precaching explicitly requires these assets to persist. They
  do however remain temporary.

  This approach leaves the legacy mode of operation (ALL_SIDES)
  unchanged in this respect.

+++ OpenSim/Region/CoreModules/Scripting/DynamicTexture/DynamicTextureModule.cs
@@ -350,8 +350,11 @@ namespace OpenSim.Region.CoreModules.Scripting.DynamicTexture
 
                 part.UpdateTexture(tmptex);
                 
-                // scene.CommsManager.AssetCache.ExpireAsset(oldID);
-                scene.AssetService.Delete(oldID.ToString());
+                if (Face == ALL_SIDES)
+                {
+                    // scene.CommsManager.AssetCache.ExpireAsset(oldID);
+                    scene.AssetService.Delete(oldID.ToString());
+                }
             }
 
             private byte[] BlendTextures(byte[] frontImage, byte[] backImage, bool setNewAlpha, byte newAlpha)
